[DSPlayer] Improved Fast Forward (thanks to Ti-Ben)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@28681 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/DSPlayer/DSPlayer.h
@@ -131,8 +131,8 @@ protected:
   virtual void OnStartup();
   virtual void OnExit();
   virtual void Process();
-  int  m_currentSpeed;
-  double m_currentRate;
+  int m_currentSpeed;
+  int m_currentRate;
   CDSGraph m_pDsGraph;
   CPlayerOptions m_PlayerOptions;
   CURL m_Filename;
